refactor: params.ChainConfigExtra -> params/extras
