+++ params/extras/config_extra_test.go
@@ -1,7 +1,7 @@
 // (c) 2024 Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package params
+package extras
 
 import (
 	"testing"
